Refactor: Moved util methods from SparkBigQueryUtil -> BigQueryUtil
