+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryUtil.java
@@ -19,8 +19,6 @@ import static com.google.cloud.bigquery.connector.common.BigQueryConfigurationUt
 import static com.google.cloud.bigquery.connector.common.BigQueryConfigurationUtil.getOptionFromMultipleParams;
 
 import com.google.cloud.bigquery.JobInfo;
-import com.google.cloud.bigquery.StandardTableDefinition;
-import com.google.cloud.bigquery.TableDefinition;
 import com.google.cloud.bigquery.TableId;
 import com.google.cloud.bigquery.TableInfo;
 import com.google.cloud.bigquery.connector.common.BigQueryConfigurationUtil;
@@ -290,34 +288,5 @@ public class SparkBigQueryUtil {
                     "dataproc_job_uuid",
                     BigQueryUtil.sanitizeLabelValue(tag.substring(tag.lastIndexOf('_') + 1))));
     return labels.build();
-  }
-
-  /**
-   * BigLake Managed tables are not represented by a dedicated type. Instead, their presence is
-   * indicated by the field 'bigLakeConfiguration' within the StandardTableDefinition {@link
-   * StandardTableDefinition#getBigLakeConfiguration()}.
-   *
-   * @param table the table to check
-   * @return Returns true if the table is a BigLake Managed table.
-   */
-  public static boolean isBigLakeManagedTable(TableInfo table) {
-    TableDefinition tableDefinition = table.getDefinition();
-    return tableDefinition.getType() == TableDefinition.Type.TABLE
-        && tableDefinition instanceof StandardTableDefinition
-        && ((StandardTableDefinition) tableDefinition).getBigLakeConfiguration() != null;
-  }
-
-  /**
-   * Since StandardTableDefinition (table_type == TableDefinition.Type.TABLE) can represent both
-   * BigQuery native tables and BigLake Managed tables, the absence of the "bigLakeConfiguration"
-   * field within the StandardTableDefinition {@link
-   * StandardTableDefinition#getBigLakeConfiguration()}. indicates a BigQuery native table.
-   *
-   * @param table the table to check
-   * @return Returns true if the table is a BigQuery Native table.
-   */
-  public static boolean isBigQueryNativeTable(TableInfo table) {
-    return table.getDefinition().getType() == TableDefinition.Type.TABLE
-        && !isBigLakeManagedTable(table);
   }
 }
